코드리뷰 피드백-1
코드간소화 진행했습니다.
